Rework ArgumentSanitizer to only one sanitizer step
This allows to properly unwrap InternalObject and ExternalObject
even when they are encapsulated into another object.
The trick is to make these object non marshalable by adding a
non marshalable member, so that dissection takes place.

However the second step was just flat and didn't recognize
encapsulated InternalObject or ExternalObject. Therefore
it is now combined into one step, with the target_event_loop as
an argument and a new approach for the "internal" detection.