Track the origination of calls to event scope methods
Until now only the called event loop and the target event loop was
considered (in Sanitizer) and registered (in WrappedObject etc.).
This was sufficient to track objects for one Eventbox instance.

However in the context of several Eventbox instances it's necessary
to know to origin event scope in order to unwrap objects which are
returned from another event scope.

To make this happen the currently executed event loop is written to
a thread local variable and is recognized at event scope method entry.

Since the origin scope is known now, all wrapped objects can be handled
equally, so that InternalObject and ExternalObject can be unified
to just WrappedObject.