Add automatic wrapping of external data for €-variables
This adds a new feature for working with external objects, which are
passed through the event scope only.
It is similar to Eventbox#shared_object, but works for any objects.
Even immutable types or BasicObject derivations can be used.

The € character is available on many keyboard layouts across the world
and is usable to prefix variables. And it's the counterpart to the
$-variables we already have in ruby.

Proper documentation is still missing, but I'll add it in a later commit.