Add a new Eventbox option: gc_actions
This also disables the previous default to shutdown action threads, when
the Eventbox instance is no longer referenced. The reason is:
Automatic GC'ing turned out to be not very reliable. In some cases running
actions permanently block objects to be GC'ed, so that they are never freed.
In other cases the object is collected unintensionally at random (like
MyConnection in test/examples/server_test.rb).
Both cases can be hard to debug, so that I think it's better to require
explicit shutdown by default.