Allow external closures to be called by async_call and async_proc
The closure is then called on the thread of the defining method or proc.
It is therefore only allowed as long as the defining call didn't return.
That also means that the defining call must be a yield_call or yield_proc.

This doesn't change the behaviour, that external blocks are called on the
thread of the calling method or proc, when the latter is a sync or yield call.

It would be desireable to always call the closure on the thread of the defining method
for async, sync and yield callers, as long as the defining method didn't return.
However this would require to change the execution thread from the defining method
to the calling method, as soon as the defining method has returned.
Unfortunately this would be a racy operation, making the switch unpreditable.
So the current solution is to define the thread by the type of the calling method.